move: feature package -> feature module
