+++ app/src/main/java/com/oreocube/booksearch/BookSearchAppState.kt
@@ -10,7 +10,7 @@ import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
 import com.oreocube.booksearch.core.navigation.TopLevelDestination
-import com.oreocube.booksearch.feature.favorite.library.navigateToFavoriteLibrary
+import com.oreocube.booksearch.feature.favorite.navigateToFavoriteLibrary
 import com.oreocube.booksearch.feature.home.navigateToHome
 
 @Stable
