+++ feature/favorite/src/main/java/com/oreocube/booksearch/feature/favorite/FavoriteLibraryViewModel.kt
@@ -1,4 +1,4 @@
-package com.oreocube.booksearch.feature.favorite.library
+package com.oreocube.booksearch.feature.favorite
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
