+++ feature/library/src/main/java/com/oreocube/booksearch/feature/library/SearchLibraryScreen.kt
@@ -24,12 +24,12 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.oreocube.booksearch.R
 import com.oreocube.booksearch.core.ui.component.BookSearchButton
 import com.oreocube.booksearch.core.ui.component.BookSearchTopBar
 import com.oreocube.booksearch.core.ui.theme.Brown20
 import com.oreocube.booksearch.core.ui.theme.Gray10
 import com.oreocube.booksearch.core.ui.theme.Gray20
+import com.oreocube.booksearch.core.ui.R
 import com.oreocube.booksearch.domain.model.Library
 
 @Composable
